htmlEscape/escapeUnsafeChars
