+++ modules/adgenerationBidAdapter.js
@@ -1,10 +1,11 @@
-import {escapeUnsafeChars, deepAccess} from '../src/utils.js';
+import {deepAccess} from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, NATIVE} from '../src/mediaTypes.js';
 import {config} from '../src/config.js';
 import {convertOrtbRequestToProprietaryNative} from '../src/native.js';
 import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
 import {tryAppendQueryString} from '../libraries/urlUtils/urlUtils.js';
+import {escapeUnsafeChars} from '../libraries/htmlEscape/htmlEscape.js';
 
 const ADG_BIDDER_CODE = 'adgeneration';
